rename 'build' to 'deploy'
